+++ zerver/lib/narrow.py
@@ -333,8 +333,24 @@ class NarrowBuilder:
         return method(query, operand, maybe_negate)
 
     def by_has(self, query: Select, operand: str, maybe_negate: ConditionTransform) -> Select:
-        if operand not in ["attachment", "image", "link"]:
+        if operand not in ["attachment", "image", "link", "reaction"]:
             raise BadNarrowOperatorError("unknown 'has' operand " + operand)
+
+        if operand == "reaction":
+            if self.msg_id_column.name == "message_id":
+                # If the initial query uses `zerver_usermessage`
+                check_col = literal_column("zerver_usermessage.message_id", Integer)
+            else:
+                # If the initial query doesn't use `zerver_usermessage`
+                check_col = literal_column("zerver_message.id", Integer)
+            exists_cond = (
+                select([1])
+                .select_from(table("zerver_reaction"))
+                .where(check_col == literal_column("zerver_reaction.message_id", Integer))
+                .exists()
+            )
+            return query.where(maybe_negate(exists_cond))
+
         col_name = "has_" + operand
         cond = column(col_name, Boolean)
         return query.where(maybe_negate(cond))